# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hosaka Appliance",
    "description": "API for the Hosaka Field Terminal Pi. The `/api/v1/*` surface is the public, versioned contract that hosakactl (Mac/Linux client) and the kiosk SPA both consume. See the published reference at https://<your-gh-user>.github.io/Hosaka/ for static docs.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/inbox/events": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Append-only inbox / notification feed",
        "operationId": "v1_inbox_events_api_v1_inbox_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 200,
              "title": "Limit"
            }
          },
          {
            "name": "topic",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Topic"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxFeedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "v1"
        ],
        "summary": "Append a notification event",
        "operationId": "v1_inbox_post_api_v1_inbox_events_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboxPostIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InboxEventOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events",
          "event store"
        ],
        "summary": "List Events",
        "operationId": "list_events_api_v1_events_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO8601 UTC; only events newer than this",
              "title": "Since"
            },
            "description": "ISO8601 UTC; only events newer than this"
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "function|module|command|keystroke|ui|crash|health|recovery|custom",
              "title": "Kind"
            },
            "description": "function|module|command|keystroke|ui|crash|health|recovery|custom"
          },
          {
            "name": "feature_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature Id"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "debug|info|warn|error|fatal",
              "title": "Level"
            },
            "description": "debug|info|warn|error|fatal"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 200,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-feature-id": "obs.events"
      }
    },
    "/api/v1/events/summary": {
      "get": {
        "tags": [
          "events",
          "event store"
        ],
        "summary": "Summary",
        "operationId": "summary_api_v1_events_summary_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "window_hours",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 24.0,
              "exclusiveMinimum": 0.0,
              "default": 1.0,
              "title": "Window Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-feature-id": "obs.events"
      }
    },
    "/api/v1/events/silent": {
      "get": {
        "tags": [
          "events",
          "event store"
        ],
        "summary": "Silent",
        "operationId": "silent_api_v1_events_silent_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "minutes",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1440,
              "minimum": 1,
              "default": 15,
              "title": "Minutes"
            }
          },
          {
            "name": "expected",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Comma-separated feature_ids that should be alive. If omitted, we use every feature_id seen in the last 24h as the baseline.",
              "title": "Expected"
            },
            "description": "Comma-separated feature_ids that should be alive. If omitted, we use every feature_id seen in the last 24h as the baseline."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SilentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-feature-id": "obs.events"
      }
    },
    "/api/v1/events/stats": {
      "get": {
        "tags": [
          "events",
          "event store"
        ],
        "summary": "Stats",
        "operationId": "stats_api_v1_events_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "x-feature-id": "obs.events"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts_utc": {
            "type": "string",
            "title": "Ts Utc"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "correlation_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Correlation Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "feature_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "level": {
            "type": "string",
            "title": "Level"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "payload": {
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ts_utc",
          "session_id",
          "kind",
          "source",
          "level"
        ],
        "title": "EventOut"
      },
      "EventsListOut": {
        "properties": {
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/EventOut"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "count",
          "events"
        ],
        "title": "EventsListOut"
      },
      "FeatureSummary": {
        "properties": {
          "feature_id": {
            "type": "string",
            "title": "Feature Id"
          },
          "last_seen_utc": {
            "type": "string",
            "title": "Last Seen Utc"
          },
          "events_total": {
            "type": "integer",
            "title": "Events Total"
          },
          "errors_1h": {
            "type": "integer",
            "title": "Errors 1H"
          },
          "p95_duration_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "P95 Duration Ms"
          }
        },
        "type": "object",
        "required": [
          "feature_id",
          "last_seen_utc",
          "events_total",
          "errors_1h"
        ],
        "title": "FeatureSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InboxEventOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "at": {
            "type": "integer",
            "title": "At"
          },
          "kind": {
            "type": "string",
            "enum": [
              "notify",
              "ack"
            ],
            "title": "Kind"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "topic": {
            "type": "string",
            "title": "Topic"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "success",
              "warn",
              "error"
            ],
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "target": {
            "type": "string",
            "title": "Target"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "event_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Ref"
          },
          "prev_hash": {
            "type": "string",
            "title": "Prev Hash"
          },
          "hash": {
            "type": "string",
            "title": "Hash"
          },
          "acked": {
            "type": "boolean",
            "title": "Acked",
            "default": false
          },
          "ack_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ack At"
          },
          "ack_author": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ack Author"
          }
        },
        "type": "object",
        "required": [
          "id",
          "at",
          "kind",
          "author",
          "node_id",
          "topic",
          "severity",
          "title",
          "body",
          "target",
          "tags",
          "prev_hash",
          "hash"
        ],
        "title": "InboxEventOut"
      },
      "InboxFeedOut": {
        "properties": {
          "chain_head": {
            "type": "string",
            "title": "Chain Head"
          },
          "chain_ok": {
            "type": "boolean",
            "title": "Chain Ok"
          },
          "node_id": {
            "type": "string",
            "title": "Node Id"
          },
          "notifications": {
            "items": {
              "$ref": "#/components/schemas/InboxEventOut"
            },
            "type": "array",
            "title": "Notifications"
          }
        },
        "type": "object",
        "required": [
          "chain_head",
          "chain_ok",
          "node_id",
          "notifications"
        ],
        "title": "InboxFeedOut"
      },
      "InboxPostIn": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 140,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 4000,
            "title": "Body",
            "default": ""
          },
          "author": {
            "type": "string",
            "maxLength": 64,
            "title": "Author",
            "default": "operator"
          },
          "topic": {
            "type": "string",
            "maxLength": 64,
            "title": "Topic",
            "default": "general"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "success",
              "warn",
              "error"
            ],
            "title": "Severity",
            "default": "info"
          },
          "target": {
            "type": "string",
            "maxLength": 128,
            "title": "Target",
            "default": "broadcast"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "InboxPostIn"
      },
...
      "SilentOut": {
        "properties": {
          "silent_for_minutes": {
            "type": "integer",
            "title": "Silent For Minutes"
          },
          "features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "silent_for_minutes",
          "features"
        ],
        "title": "SilentOut"
      },
      "StatsOut": {
        "properties": {
          "db_exists": {
            "type": "boolean",
            "title": "Db Exists"
          },
          "db_path": {
            "type": "string",
            "title": "Db Path"
          },
          "counters": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counters"
          }
        },
        "type": "object",
        "required": [
          "db_exists",
          "db_path",
          "counters"
        ],
        "title": "StatsOut"
      },
      "SummaryOut": {
        "properties": {
          "window_hours": {
            "type": "number",
            "title": "Window Hours"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/FeatureSummary"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "required": [
          "window_hours",
          "features"
        ],
        "title": "SummaryOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}